refactor(cli): extract manifest runtime state domain
